% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{bw}
\alias{bw}
\title{Baum Welch Alg}
\usage{
bw(p0, e0, v0, tol)
}
\arguments{
\item{p0}{prob matrix from uniform dist}

\item{e0}{emissions matrix, top row is true, rest is from uniform dist}

\item{v0}{starting point, uniform dist}

\item{tol}{cutoff for convergence}
}
\value{
out list comprised of p0,e0, and v0 after converging
}
\description{
Baum Welch Alg
}
