% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{em}
\alias{em}
\title{EM alg}
\usage{
em(n, pa, pb, po)
}
\arguments{
\item{n}{vector of na, nab, nb, no}

\item{pa}{probability of type A}

\item{pb}{probability of type B}

\item{po}{probability of type O}
}
\value{
3x10 matrix, each row corresponds to pa, pb, and po
}
\description{
EM alg
}
