% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.R
\name{logreg}
\alias{logreg}
\title{Calculate log regression coefficients}
\usage{
logreg(x, y, method)
}
\arguments{
\item{x}{matrix of data points by parameter number}

\item{y}{reponse variable}

\item{method}{1 if grad desc, 2 if newton}
}
\value{
coefficients, confidence interval (width/2), likelihood over iterations
}
\description{
Calculate log regression coefficients
}
