% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matmul.R
\name{matmul}
\alias{matmul}
\title{This function takes in two matrices and a vector and multiplies them either (AB)x or A(Bx)}
\usage{
matmul(A, B, x, arg)
}
\arguments{
\item{A}{First input matrix, square}

\item{B}{Second input matrix, square}

\item{x}{Vector input, assume dimension matches that of the matrices}

\item{arg}{Input of 1 or 2 corresponding to order of multiplication above}
}
\value{
Product of the matrices and vector
}
\description{
This function takes in two matrices and a vector and multiplies them either (AB)x or A(Bx)
}
\examples{
matmul(matrix(1:9,nrow=3),matrix(10:18,nrow=3),matrix(1:3,nrow=3),1)
}
