% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_qr.R
\name{ols_qr}
\alias{ols_qr}
\title{This function uses the QR method to estimate OLS regresion coefficients}
\usage{
ols_qr(data)
}
\arguments{
\item{data}{CSV file where first column is the response vector}
}
\value{
vector of coefficients
}
\description{
This function uses the QR method to estimate OLS regresion coefficients
}
\examples{
ols_qr(data)
}
