% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_svd.R
\name{ols_svd}
\alias{ols_svd}
\title{This function uses the SVD decomposition to estimate OLS regresion coefficients}
\usage{
ols_svd(data)
}
\arguments{
\item{data}{CSV file where first column is the response vector}
}
\value{
vector of coefficients
}
\description{
This function uses the SVD decomposition to estimate OLS regresion coefficients
}
\examples{
ols_svd(data)
}
