% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posdef.R
\name{Posdef}
\alias{Posdef}
\title{Generating a random positive-definite matrix with user-specified positive eigenvalues}
\usage{
Posdef(n, ev = runif(n, 0, 10))
}
\arguments{
\item{n}{First input, dimension of matrix}
}
\value{
Random pos def matrix
}
\description{
Generating a random positive-definite matrix with user-specified positive eigenvalues
}
\examples{
Posdef(4)
}
