% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Calculate sigmoid}
\usage{
sigmoid(theta, x)
}
\arguments{
\item{theta}{vector of length equal to number of parameters}

\item{x}{matrix of data points by parameter number}
}
\value{
sigmoid of each data point
}
\description{
Calculate sigmoid
}
